Change skip move gesture to "Tap->Swipe"
Before it was "Swipe->Tap" which meant we had to wait once swiped to see
if it is followed by a tap. In order to reduce this latency, swap the
order: first tap then swipe is "skip", only swipe is a normal "pass on" move.
This way, once there is a swipe event, we can directly send the action
to the server and don't have to wait.