Refactor frontend & add restart/reset functionality
In the frontend, there is now a layer manager, responsible to handle the
z indices when requesting to show something on top. This greatly simplified
a lot of code. Also put more code into individual files, e.g. flickering effect.
Restructured index.html, so that z-index is not needed in the beginning.

There is now a button "Reset player order (optional)" in the admin dashboard
to redo the calibration phase. Added logic for this in the backend.

Also, we can now restart the game without all clients having to reconnect.
No new logic was needed for this in the backend.